+++ app/mypage/components/Party/index.tsx
@@ -1,3 +1,4 @@
 import PartyItem from './PartyItem';
+import PartyList from './PartyList';
 
-export { PartyItem };
+export { PartyItem, PartyList };
