+++ app/mypage/components/UserInfo/UserInfo.tsx
@@ -10,7 +10,7 @@ export default async function UserInfo({ queryClient }: { queryClient: QueryClie
   return (
     <section className="flex items-center px-[30px] justify-between h-[200px] shadow-md rounded-xl bg-white pb-7">
       <div className="flex items-center">
-        <Avatar src={`${process.env.NEXT_PUBLIC_IMAGE_URL + user_image_url}`} className="w-[100px] h-[100px] mr-4" />
+        <Avatar src={`${user_image_url}`} className="w-[100px] h-[100px] mr-4" />
 
         <div className="flex flex-col gap-3">
           <p className="text-3xl font-semibold">{user_name}</p>
