+++ app/mypage/page.tsx
@@ -3,26 +3,10 @@
 import { useEffect } from 'react';
 import { useRouter } from 'next/navigation';
 import AddButon from '@/components/common/Button/AddButton';
-import Slider from 'react-slick';
 import { getCookie } from 'cookies-next';
 import { ACCESS_TOKEN } from '@/constants';
 import { dehydrate, HydrationBoundary, useQueryClient } from '@tanstack/react-query';
-import { PartyItem, CongressmanList, UserInfo } from './components';
-
-const partyList = [
-  {
-    label: '더불어민주당',
-    src: '/mock/party/더불어민주당_로고.png',
-  },
-  {
-    label: '국민의힘',
-    src: '/mock/party/국민의힘_로고.png',
-  },
-  {
-    label: '정의당',
-    src: '/mock/party/정의당_로고.png',
-  },
-];
+import { CongressmanList, UserInfo, PartyList } from './components';
 
 const congressmanList = [
   {
@@ -111,14 +95,6 @@ export default function MyPage() {
   const router = useRouter();
   const accessToken = getCookie(ACCESS_TOKEN);
   const queryClient = useQueryClient();
-  const sliderSettings = {
-    arrows: false,
-    dots: false,
-    slidesToScroll: 1,
-    swipeToSlide: true,
-    speed: 2000,
-    variableWidth: true,
-  };
 
   useEffect(() => {
     if (!accessToken) {
@@ -133,21 +109,7 @@ export default function MyPage() {
       <div className="flex flex-col h-full gap-8">
         <UserInfo queryClient={queryClient} />
 
-        <section className="pl-[30px] flex flex-col gap-6">
-          <div className="flex items-center justify-between pr-[30px]">
-            <p className="text-xl font-semibold">
-              팔로우한 정당 &middot;<span className="text-[#555555]"> 3</span>
-            </p>
-
-            <AddButon />
-          </div>
-
-          <Slider {...sliderSettings}>
-            {partyList.map((party) => (
-              <PartyItem key={party.label} {...party} />
-            ))}
-          </Slider>
-        </section>
+        <PartyList qeuryClient={queryClient} />
 
         <hr className="mx-[30px] bg-[#E0E0E0]" />
