+++ types/type/user/followingParty.ts
@@ -1,5 +1,7 @@
-export interface FollowingPartyResponse {
+export interface FollowingPartyType {
   party_id: number;
   party_name: string;
   party_image_url: string;
 }
+
+export type FollowingPartyResponse = FollowingPartyType[];
